requirements.txt
openai
pymongo
requests
bs4
python-dotenv
flask